Uploaded Critical Path Finder (CPM) file.
Problem: You have a project consisting of various activities that need to be completed. Each activity has a duration, and there can be dependencies between activities. The DataFrame contains three columns:



"Activities": This column represents the names or identifiers of the activities.

"Predecessor": This column specifies the activities that need to be completed before a particular activity can start.
It can contain multiple predecessors separated by commas.

"Durations": This column represents the duration (in some unit of time) required to complete each activity.
You need to find the critical path of the project, which is the sequence of activities that has the longest total duration.
The critical path determines the minimum time required to complete the entire project.
Additionally, you need to calculate the minimum durations to complete all the activities.